refactor(identity): split server.go into domain endpoint files
Extract RPC handlers from server.go (987 lines) into focused files:
- grpc_identity_endpoints.go: CreateIdentity, RetrieveIdentity, UpdateIdentity, ListIdentities, SuspendIdentity, ReactivateIdentity
- grpc_auth_endpoints.go: Authenticate, SetPassword, ChangePassword, RequestPasswordReset, CompletePasswordReset
- grpc_role_endpoints.go: GrantRole, RevokeRole, ListRoleAssignments
- grpc_invitation_endpoints.go: InviteUser, AcceptInvitation

server.go retains Service struct, constructor, options, health check, and shared helpers.
Removes identity server.go from knownOversizedFiles (now 217 lines, well under 800 limit).
Pure code motion - no behavioral changes.
