+++ tests/architecture/size_test.go
@@ -26,7 +26,6 @@ const (
 // Do NOT add new entries — split large files instead.
 var knownOversizedFiles = map[string]bool{
 	"services/current-account/service/saga_handlers.go":          true,
-	"services/identity/service/server.go":                        true,
 	"services/market-information/service/observation_service.go": true,
 	"services/mcp-server/internal/tools/economy.go":              true,
 	"services/mcp-server/internal/tools/refdata.go":              true,
